Don't consider no-child-process sub-command error as such in some cases.
Due to a race-condition in Golang, it's possible that a cmd.Start() may result
in a "no child process" error even though the child process was created and in
fact succeeded. See https://github.com/kubernetes/kubernetes/issues/103753 for
full context.

This change causes the code to ignore "no child process" errors if the
sub-command process succeeds.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>